Separate one function from index.ts
